Using Redis to cache the SMS code and dish data.

+++ src/main/java/org/example/radiation/controller/DishController.java
@@ -15,9 +15,13 @@ import org.example.radiation.service.DishFlavorService;
 import org.example.radiation.service.DishService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 @RestController
@@ -33,6 +37,9 @@ public class DishController {
     @Autowired
     private CategoryService categoryService;
 
+    @Autowired
+    private RedisTemplate redisTemplate;
+
     /**
      * 新增菜品
      * @param dishDto
@@ -43,6 +50,10 @@ public class DishController {
         log.info(dishDto.toString());
 
         dishService.saveWithFlavor(dishDto);
+
+        // 清理菜品对应分类的redis缓存数据
+        String key = "dish_" + dishDto.getCategoryId() + "_1";
+        redisTemplate.delete(key);
 
         return R.success("新增菜品成功");
     }
@@ -121,6 +132,10 @@ public class DishController {
 
         dishService.updateWithFlavor(dishDto);
 
+        // 清理菜品对应分类的redis缓存数据
+        String key = "dish_" + dishDto.getCategoryId() + "_1";
+        redisTemplate.delete(key);
+
         return R.success("修改菜品成功");
     }
 
@@ -132,15 +147,29 @@ public class DishController {
      */
     @PostMapping("/status/{status}")
     public R<String> status(@PathVariable Integer status, @RequestParam List<Long> ids){
-        // 条件构造器
-        LambdaUpdateWrapper<Dish> queryWrapper = new LambdaUpdateWrapper<>();
+        // 清理所有菜品redis缓存
+        //Set keys = redisTemplate.keys("dish_*");
+        //redisTemplate.delete(keys);
 
-        // 添加条件
-        queryWrapper.in(ids != null && ids.size() > 0, Dish::getId, ids);
-        queryWrapper.set(status != null, Dish::getStatus, status);
+        // 清除菜品对应分类在redis中的缓存
+        LambdaQueryWrapper<Dish> queryWrapper = new LambdaQueryWrapper<>();
+        queryWrapper.in(Dish::getId, ids);
+        List<Dish> categoryIdList = dishService.list(queryWrapper);
+
+        // 将查询到的菜品所属分类id去重
+        TreeSet<Long> categoryIdSet = categoryIdList.stream().map(Dish::getCategoryId).collect(Collectors.toCollection(TreeSet::new));
+        for(Long id : categoryIdSet){
+            String key = "dish_" + String.valueOf(id) + "_1";
+            redisTemplate.delete(key);
+        }
+
+        List<Dish> list = categoryIdList.stream().map((item) -> {
+            item.setStatus(status);
+            return item;
+        }).collect(Collectors.toList());
 
         // 更新数据库
-        dishService.update(queryWrapper);
+        dishService.updateBatchById(list);
 
         return R.success("菜品信息修改成功");
     }
@@ -151,7 +180,18 @@ public class DishController {
      */
     @DeleteMapping
     public R<String> delete(@RequestParam List<Long> ids){
+        // 查询dish表，判断是否含有在售菜品
+        LambdaQueryWrapper<Dish> queryWrapper = new LambdaQueryWrapper<>();
+        queryWrapper.in(Dish::getId, ids);
+        queryWrapper.eq(Dish::getStatus, 1);
+        int count = dishService.count(queryWrapper);
 
+        // 如果包含在售菜品，则拒绝删除
+        if(count > 0){
+            return R.error("所选包含在售菜品，无法删除");
+        }
+
+        // 否则正常删除
         dishService.deleteWithFlavor(ids);
 
         return R.success("删除菜品成功");
@@ -164,7 +204,18 @@ public class DishController {
      */
     @GetMapping("/list")
     public R<List<DishDto>> list(Dish dish){
-        // 构建查询条件对象
+        List<DishDto> dishDtoList = null;
+        String key = "dish_" + dish.getCategoryId() + "_" + dish.getStatus();
+
+        // 先从redis中获取缓存数据
+        dishDtoList =(List<DishDto>) redisTemplate.opsForValue().get(key);
+
+        // 如果存在，直接返回
+        if(dishDtoList != null) {
+            return R.success(dishDtoList);
+        }
+
+        // 如果不存在，查询数据库
         LambdaQueryWrapper<Dish> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.eq(dish.getCategoryId() != null, Dish::getCategoryId, dish.getCategoryId());
         queryWrapper.eq(dish.getName() != null, Dish::getName, dish.getName());
@@ -175,7 +226,7 @@ public class DishController {
 
         List<Dish> list = dishService.list(queryWrapper);
 
-        List<DishDto> dishDtoList = list.stream().map((item) ->{
+        dishDtoList = list.stream().map((item) ->{
             // 对象拷贝
             DishDto dishDto = new DishDto();
             BeanUtils.copyProperties(item, dishDto);
@@ -200,6 +251,9 @@ public class DishController {
 
             return dishDto;
         }).collect(Collectors.toList());
+
+        // 如果redis中不存在，还需将查到的数据放入redis中,有效期为60分钟
+        redisTemplate.opsForValue().set(key, dishDtoList, 60, TimeUnit.MINUTES);
 
         return R.success(dishDtoList);
     }
